--- cogs/__init__.py
@@ -1,2 +1,0 @@
-# cogs/__init__.py
-# This file can be left empty or used to initialize the cogs
